clippy --fix

+++ plugin/renderer/codegen/hlsl/cgen_type/layout_c.hlsl
@@ -5,7 +5,7 @@
 
 
     struct LayoutC {
-        float a;
+        half a;
     };
 
 #endif
